+++ docknv/image/__init__.py
@@ -1,3 +1,4 @@
 """Image handler."""
 
 from .methods import *  # noqa
+from .models import *  # noqa